+++ ikarus/finiteelements/mechanics/kirchhoffloveshell.hh
@@ -16,6 +16,7 @@
 #include <ikarus/finiteelements/ferequirements.hh>
 #include <ikarus/finiteelements/feresulttypes.hh>
 #include <ikarus/finiteelements/mechanics/loads.hh>
+#include <ikarus/finiteelements/mechanics/massmatrix.hh>
 #include <ikarus/finiteelements/mechanics/membranestrains.hh>
 #include <ikarus/finiteelements/physicshelper.hh>
 #include <ikarus/utils/linearalgebrahelper.hh>
@@ -31,6 +32,7 @@ struct KirchhoffLoveShellPre
 {
   YoungsModulusAndPoissonsRatio material;
   double thickness;
+  double density;
 
   template <typename PreFE, typename FE>
   using Skill = KirchhoffLoveShell<PreFE, FE>;
@@ -99,7 +101,8 @@ public:
    */
   KirchhoffLoveShell(const Pre& pre)
       : mat_{pre.material},
-        thickness_{pre.thickness} {}
+        thickness_{pre.thickness},
+        density_{pre.density} {}
 
 protected:
   /**
@@ -179,6 +182,7 @@ private:
   // DefaultMembraneStrain membraneStrain_;
   YoungsModulusAndPoissonsRatio mat_;
   double thickness_;
+  double density_;
 
   size_t numberOfNodes_{0};
   int order_{};
@@ -238,6 +242,10 @@ protected:
   void calculateMatrixImpl(
       const Requirement& par, const MatrixAffordance& affordance, typename Traits::template MatrixType<ST> K,
       const std::optional<std::reference_wrapper<const Eigen::VectorX<ST>>>& dx = std::nullopt) const {
+    if (affordance == MatrixAffordance::linearMass) {
+      calculateMassImpl<ST>(par, affordance, K);
+      return;
+    }
     if (affordance != MatrixAffordance::stiffness)
       DUNE_THROW(Dune::NotImplemented, "MatrixAffordance not implemented: " + toString(affordance));
     using namespace Dune::DerivativeDirections;
@@ -277,6 +285,19 @@ protected:
       }
     }
     K.template triangularView<Eigen::StrictlyLower>() = K.transpose();
+  }
+
+  template <typename ScalarType>
+  void calculateMassImpl(const Requirement& par, const MatrixAffordance& affordance,
+                         typename Traits::template MatrixType<> M) const {
+    const auto geo  = underlying().localView().element().geometry();
+    const auto rhoT = thickness_ * density_;
+
+    for (const auto& [gpIndex, gp] : localBasis_.viewOverIntegrationPoints()) {
+      const auto intElement = geo.integrationElement(gp.position()) * gp.weight();
+      const auto& N         = localBasis_.evaluateFunction(gpIndex);
+      evaluateKroneckerProduct<worldDim>(intElement, N, rhoT, M);
+    }
   }
 
   template <typename ST>
@@ -428,13 +449,14 @@ protected:
 
 /**
  * \brief A struct containing information about the Youngs Modulus,
- * Poisson's ratio and the thickness for the Kirchhoff-Love shell element.
+ * Poisson's ratio, thickness and the density (defaults so 1.0) for the Kirchhoff-Love shell element.
  */
 struct KlArgs
 {
   double youngs_modulus;
   double nu;
   double thickness;
+  double density{1.0};
 };
 
 /**
@@ -443,7 +465,7 @@ struct KlArgs
  * \return A Kirchhoff-Love shell pre finite element.
  */
 auto kirchhoffLoveShell(const KlArgs& args) {
-  KirchhoffLoveShellPre pre({.emodul = args.youngs_modulus, .nu = args.nu}, args.thickness);
+  KirchhoffLoveShellPre pre({.emodul = args.youngs_modulus, .nu = args.nu}, args.thickness, args.density);
 
   return pre;
 }