+++ ikarus/finiteelements/mechanics/nonlinearelastic.hh
@@ -22,6 +22,7 @@
   #include <ikarus/finiteelements/fehelper.hh>
   #include <ikarus/finiteelements/ferequirements.hh>
   #include <ikarus/finiteelements/mechanics/loads.hh>
+  #include <ikarus/finiteelements/mechanics/massmatrix.hh>
   #include <ikarus/finiteelements/mechanics/materials/tags.hh>
   #include <ikarus/finiteelements/physicshelper.hh>
   #include <ikarus/utils/defaultfunctions.hh>
@@ -42,6 +43,7 @@ struct NonLinearElasticPre
 {
   using Material = MAT;
   MAT material;
+  double density;
 
   template <typename PreFE, typename FE>
   using Skill = NonLinearElastic<PreFE, FE, NonLinearElasticPre>;
@@ -82,7 +84,8 @@ public:
    * \param pre The pre fe
    */
   explicit NonLinearElastic(const Pre& pre)
-      : mat_{pre.material} {}
+      : mat_{pre.material},
+        density_{pre.density} {}
 
 protected:
   /**
@@ -233,6 +236,7 @@ private:
   std::shared_ptr<const Geometry> geo_;
   Dune::CachedLocalBasis<std::remove_cvref_t<LocalBasisType>> localBasis_;
   Material mat_;
+  double density_;
   size_t numberOfNodes_{0};
   int order_{};
 
@@ -248,6 +252,10 @@ protected:
   void calculateMatrixImpl(
       const Requirement& par, const MatrixAffordance& affordance, typename Traits::template MatrixType<> K,
       const std::optional<std::reference_wrapper<const Eigen::VectorX<ScalarType>>>& dx = std::nullopt) const {
+    if (affordance == MatrixAffordance::linearMass) {
+      calculateMassImpl<ScalarType>(par, affordance, K);
+      return;
+    }
     using namespace Dune::DerivativeDirections;
     using namespace Dune;
     const auto uFunction = displacementFunction(par, dx);
@@ -267,6 +275,16 @@ protected:
           K.template block<myDim, myDim>(i * myDim, j * myDim) += (bopI.transpose() * C * bopJ + kgIJ) * intElement;
         }
       }
+    }
+  }
+
+  template <typename ScalarType>
+  void calculateMassImpl(const Requirement& par, const MatrixAffordance& affordance,
+                         typename Traits::template MatrixType<> M) const {
+    for (const auto& [gpIndex, gp] : localBasis_.viewOverIntegrationPoints()) {
+      const auto intElement = geo_->integrationElement(gp.position()) * gp.weight();
+      const auto& N         = localBasis_.evaluateFunction(gpIndex);
+      evaluateKroneckerProduct<myDim>(intElement, N, density_, M);
     }
   }
 
@@ -315,11 +333,12 @@ protected:
  * \brief A helper function to create a non-linear elastic pre finite element.
  * \tparam MAT Type of the material.
  * \param mat Material parameters for the non-linear elastic element.
+ * \param density Density of non-linear elastic element (defaults to 1.0)
  * \return A non-linear elastic pre finite element.
  */
 template <typename MAT>
-auto nonLinearElastic(const MAT& mat) {
-  NonLinearElasticPre<MAT> pre(mat);
+auto nonLinearElastic(const MAT& mat, double density = 1.0) {
+  NonLinearElasticPre<MAT> pre(mat, density);
 
   return pre;
 }